requirements_osint.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
linkedin-api>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.29.0
pandas>=2.1.0
plotly>=5.18.0
python-multipart>=0.0.6
fastapi>=0.104.0
uvicorn>=0.24.0 